Creacion de unit test para yacc
