--- tests/test_answers.txt
@@ -1,12 +1,0 @@
-tests/test_1.txt
-01 | ENDOF - - -\n
-tests/test_2.txt
-01 | ENDOF - - -\n02 | EQ 1 - a\n03 | ENDOF - - -\n
-tests/test_4.txt
-01 | ENDOF - - -\n02 | EQ 1 - a\n03 | ENDOF - - -\n04 | EQ 10 - x\n05 | EQ 2 - y\n06 | MUL x y T6\n07 | EQ T6 - a\n08 | ADD x y T8\n09 | EQ T8 - b\n10 | DIV x y T10\n11 | EQ T10 - c\n12 | SUB x y T12\n13 | EQ T12 - d\n14 | ENDOF - - -\n
-tests/test_5.txt
-01 | ENDOF - - -\n02 | EQ 1 - a\n03 | ENDOF - - -\n04 | EQ 10 - x\n05 | EQ 2 - y\n06 | MUL x y T6\n07 | EQ T6 - a\n08 | ADD x y T8\n09 | EQ T8 - b\n10 | DIV x y T10\n11 | EQ T10 - c\n12 | SUB x y T12\n13 | EQ T12 - d\n14 | ENDOF - - -\n15 | EQ 1 - a\n16 | EQ 2 - b\n17 | EQ 0 - c\n18 | GT a b T18\n19 | GOTOF T18 - 23\n20 | ADD a 1 T20\n21 | EQ T20 - c\n22 | GOTO - - 30\n23 | BEQ b a T23\n24 | GOTOF T23 - 28\n25 | ADD b 1 T25\n26 | EQ T25 - c\n27 | GOTO - - 30\n28 | ADD 1 2 T28\n29 | EQ T28 - c\n30 | ENDOF - - -\n
-tests/test_6.txt
-01 | ENDOF - - -\n02 | EQ 1 - a\n03 | ENDOF - - -\n04 | EQ 10 - x\n05 | EQ 2 - y\n06 | MUL x y T6\n07 | EQ T6 - a\n08 | ADD x y T8\n09 | EQ T8 - b\n10 | DIV x y T10\n11 | EQ T10 - c\n12 | SUB x y T12\n13 | EQ T12 - d\n14 | ENDOF - - -\n15 | EQ 1 - a\n16 | EQ 2 - b\n17 | EQ 0 - c\n18 | GT a b T18\n19 | GOTOF T18 - 23\n20 | ADD a 1 T20\n21 | EQ T20 - c\n22 | GOTO - - 30\n23 | BEQ b a T23\n24 | GOTOF T23 - 28\n25 | ADD b 1 T25\n26 | EQ T25 - c\n27 | GOTO - - 30\n28 | ADD 1 2 T28\n29 | EQ T28 - c\n30 | ENDOF - - -\n31 | EQ 1 - a\n32 | EQ 2 - b\n33 | EQ 0 - c\n34 | GT a b T34\n35 | GOTOF T34 - 39\n36 | ADD a b T36\n37 | EQ T36 - c\n38 | GOTO - - 34\n39 | ENDOF - - -\n
-tests/test_7.txt
-01 | ENDOF - - -\n02 | EQ 1 - a\n03 | ENDOF - - -\n04 | EQ 10 - x\n05 | EQ 2 - y\n06 | MUL x y T6\n07 | EQ T6 - a\n08 | ADD x y T8\n09 | EQ T8 - b\n10 | DIV x y T10\n11 | EQ T10 - c\n12 | SUB x y T12\n13 | EQ T12 - d\n14 | ENDOF - - -\n15 | EQ 1 - a\n16 | EQ 2 - b\n17 | EQ 0 - c\n18 | GT a b T18\n19 | GOTOF T18 - 23\n20 | ADD a 1 T20\n21 | EQ T20 - c\n22 | GOTO - - 30\n23 | BEQ b a T23\n24 | GOTOF T23 - 28\n25 | ADD b 1 T25\n26 | EQ T25 - c\n27 | GOTO - - 30\n28 | ADD 1 2 T28\n29 | EQ T28 - c\n30 | ENDOF - - -\n31 | EQ 1 - a\n32 | EQ 2 - b\n33 | EQ 0 - c\n34 | GT a b T34\n35 | GOTOF T34 - 39\n36 | ADD a b T36\n37 | EQ T36 - c\n38 | GOTO - - 34\n39 | ENDOF - - -\n40 | EQ 10 - b\n41 | EQ 0 - c\n42 | EQ 0 - a\n43 | LT a b T43\n44 | GOTOF T43 - 50\n45 | ADD a 1 T45\n46 | EQ T45 - a\n47 | ADD c 1 T47\n48 | EQ T47 - c\n49 | GOTO - - 43\n50 | ADD b c T50\n51 | EQ T50 - b\n52 | ENDOF - - -\n
